+++ package.json
@@ -23,7 +23,7 @@
     "cookie-parser": "^1.4.3",
     "cors": "^2.8.3",
     "ejs": "^2.5.6",
-    "eslint": "^3.19.0",
+    "eslint": "^4.0.0",
     "express": "^4.15.2",
     "express-session": "^1.15.1",
     "file": "^0.2.2",
@@ -47,10 +47,9 @@
     "react-dropzone": "^3.13.2",
     "react-redux": "^5.0.5",
     "react-router": "3.0.2",
-    "react-router-redux": "^4.1.8",
+    "react-router-redux": "^4.0.8",
     "react-simpletabs": "^0.7.0",
     "react-tabs": "^1.0.0",
-    "react-router-redux": "^4.0.8",
     "react-tournament-bracket": "^0.1.0",
     "redis": "^2.7.1",
     "redux": "^3.6.0",
